Support for Big SSD
-Tested with up to 2TB disk size
-Added support for Delay mode:
    1.real - backwards compatible, same delay that was
      implemented until now
    2.emulated - no delay, execution of test continues normally,
      delay is only logged.
-All relevant variable in code were updated to the appropriate kind,
 size of variable was increased to accommodate larger disk, and
 signed/unsigned was changed depending on actual use of variable
-Calculation of ssd_stats: utilization and read/write speed was fixed
 to account for multiple rt_analyzers
-Added ssd_stats: occupied_pages, logical_write_count,
 block_erase_count, channel_switch_to_write, channel_switch_to_read
-Added mechanism for syncing to log without arbitrary time sleep
-Updated GC: Victim list is now sorted by valid_page_nb, accelerating
 GC algorithm for large disk sizes.
-Update legacy tests for new changes
-ssd_write_read test now has tests for checking all ssd_stats
 during test of read, write, mixed read/write and GC tests

+++ eVSSIM/FTL_SOURCE/PAGE_MAP/ftl.c
@@ -123,7 +123,7 @@ TODO: review statistics implementation
 }
 
 void FTL_RESET_STATS(void){
-	int i,j;
+	uint64_t i,j;
 
 	for(i=0; i < SUPPORTED_OPERATION; i++){
 		for(j=0;j<PAGE_MAPPING_ENTRY_NB;j++){
@@ -287,7 +287,7 @@ void FTL_RECORD_STATISTICS(void){
 			fprintf(fp , "%u," , CALC_PLANE(address));
 		}
 		if (STAT_SCOPE & COLLECT_BLOCK){
-			fprintf(fp , "%u,", CALC_BLOCK(address));
+			fprintf(fp , "%lu,", CALC_BLOCK(address));
 		}
 		if (STAT_SCOPE & COLLECT_PAGE){
 			fprintf(fp , "%d," , address);
